# docs/api.yaml
...
swagger: '2.0'
info:
  title: Runners API
  description: Application for the paleo
  version: "1.0.0"
# the domain of the service
host: localhost:8080
...
schemes:
  - http
# will be prefixed to all paths
basePath: /api
produces:
  - application/json
securityDefinitions:
 AccessToken:
  description: |
   For accessing the API a valid Access token must be passed in all the queries in
   the 'Authorization' header.

   The following syntax must be used in the 'Authorization' header :

       xxxxxx.yyyyyyy.zzzzzz
  type: apiKey
  name: X-ACCESS-TOKEN
  in: header
paths:
...
  /cars/{id}:
    delete:
      summary: Delete a car
      tags:
        - Vehicles
      security:
        - AccessToken: []
      parameters:
        - name: id
          required: true
          in: path
          type: number
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Vehicle'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    patch:
      summary: Update a car
      tags:
        - Vehicles
      security:
        - AccessToken: []
      parameters:
        - name: id
          required: true
          in: path
          type: number
        - name: car
          required: true
          in: body
          schema:
            $ref: '#/definitions/Vehicle'

      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Vehicle'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    get:
      summary: Retrieve a specific car
      tags:
        - Vehicles
      security:
        - AccessToken: []
      parameters:
        - name: id
          required: true
          in: path
          type: number
      responses:
        200:
          description: An car
          schema:
            $ref: '#/definitions/Vehicle'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /cars:
    post:
      summary: Create a new car
      description:  Create a car. You can easily create a car with only a car_type, if needed.
      tags:
        - Vehicles
      security:
        - AccessToken: []
      parameters:
        - name: car
          required: true
          in: body
          schema:
            $ref: '#/definitions/Vehicle'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Vehicle'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    get:
      summary: Vehicles Types
      description:
        List all vehicles
      parameters:
        - name: status
          in: query
          description: "Allows you to retrieve a list of vehicles according to their status : [taken, free]. /vehicles?status=taken"
          required: false
          type: string
        - name: type
          in: query
          description: "Allows you to retrieve a list of vehicles according to type such as : [Vito, Limo]. /vehicles?staus=taken&type=Limo"
          required: false
          type: string
        - name: include
          in: query
          description: "Add user object if the vehicle is taken : /vehicles?include=user"
          type: string
      tags:
        - Vehicles
      security:
        - AccessToken: []
      responses:
        200:
          description: An array of vehicles
          schema:
            type: array
            items:
              $ref: '#/definitions/Vehicle'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Vehicle:
    type: object
    properties:
      id:
        description: Unique identifier (PK)
        type: integer
      name:
        description: Name of the car
        type: string
      type:
        $ref: '#/definitions/VehicleType'
      plate_number:
        description: Car plate number
        type: string
      nb_place:
        description: Number of places available in the car (driver's place is excluded)
        type: integer
      # user:
      #   $ref: '#/definitions/User'
  User:
    type: object
    properties:
      id:
        description: Unique identifier (PK)
        type: integer
      firstname:
        description: Driver's first name
        type: string
      lastname:
        description: Driver's last name
        type: string
      phone_number:
        description: Driver's phone number
        type: string
  VehicleType:
    type: object
    properties:
      id:
        type: number
      type:
        type: string
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      ui_message:
        type: string
...